Fixed Register
Had to make a register activity since I had not done any routing between login and register.
Now has some checks for password and email and a window that confirms a registration being sent.
