+++ app/src/main/java/com/example/football_manager/RegisterScreenActivity.kt
@@ -1,0 +1,27 @@
+package com.example.football_manager
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.ui.Modifier
+import com.example.football_manager.ui.theme.FootballmanagerTheme
+
+class RegisterScreenActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            FootballmanagerTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colors.background
+                ) {
+                    RegisterScreen()
+                }
+            }
+        }
+    }
+}
